Update macOS example
