+++ Examples/Example-macOS/Sources/ShuffleEmoticonViewController.swift
@@ -26,7 +26,8 @@ final class ShuffleEmoticonViewController: NSViewController {
     override func awakeFromNib() {
         super.awakeFromNib()
 
-        collectionView.register(EmoticonCollectionViewItem.self, forItemWithIdentifier: EmoticonCollectionViewItem.itemIdentifier)
+        tableView.selectionHighlightStyle = .none
+        collectionView.register(ShuffleEmoticonCollectionViewItem.self, forItemWithIdentifier: ShuffleEmoticonCollectionViewItem.itemIdentifier)
     }
 }
 
@@ -40,7 +41,7 @@ extension ShuffleEmoticonViewController: NSCollectionViewDataSource {
     }
 
     func collectionView(_ collectionView: NSCollectionView, itemForRepresentedObjectAt indexPath: IndexPath) -> NSCollectionViewItem {
-        let item = collectionView.makeItem(withIdentifier: EmoticonCollectionViewItem.itemIdentifier, for: indexPath) as! EmoticonCollectionViewItem
+        let item = collectionView.makeItem(withIdentifier: ShuffleEmoticonCollectionViewItem.itemIdentifier, for: indexPath) as! ShuffleEmoticonCollectionViewItem
         item.emoticon = data[indexPath.item]
         return item
     }
@@ -61,40 +62,5 @@ extension ShuffleEmoticonViewController: NSTableViewDataSource, NSTableViewDeleg
 private extension NSTableCellView {
     static var itemIdentifier: NSUserInterfaceItemIdentifier {
         return NSUserInterfaceItemIdentifier(String(describing: self))
-    }
-}
-
-private final class EmoticonCollectionViewItem: NSCollectionViewItem {
-    static var itemIdentifier: NSUserInterfaceItemIdentifier {
-        return NSUserInterfaceItemIdentifier(String(describing: self))
-    }
-
-    var emoticon: String {
-        get { return _textField.stringValue }
-        set { _textField.stringValue = newValue }
-    }
-
-    private let _textField = NSTextField()
-
-    override func loadView() {
-        view = NSView(frame: NSRect(x: 0, y: 0, width: 60, height: 54))
-    }
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        _textField.font = .systemFont(ofSize: 40)
-        _textField.alignment = .center
-        _textField.isEditable = false
-
-        _textField.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(_textField)
-
-        let constraints = [
-            _textField.topAnchor.constraint(equalTo: view.topAnchor),
-            _textField.bottomAnchor.constraint(equalTo: view.bottomAnchor),
-            _textField.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            _textField.trailingAnchor.constraint(equalTo: view.trailingAnchor)
-        ]
-        NSLayoutConstraint.activate(constraints)
     }
 }
