+++ Examples/Example-macOS/Sources/StringExtensions.swift
